add support for 'package' key in manifest files
one step closer to feature-parity with jake-based buildsystem. One
manifest file might contain package entries for all the manifest objects
provided in the file. If these are the same, this entry will be used
instead of the default package name. This way, a module can generate
multiple manifest files for distribution.