enable livereload for multiple instances of grunt watch
the idea is to check if the livereload port is already in use. If so,
watch task will be configured not to use livereload. Instead, there will
be a custom task 'send_livereload' that will manually trigger a
livereload event at the server running on the port. If the port is still
available, watch will be configured to start the livereload server.