Fixed: Bug 31956 - [dev] Unable to build new themes
in order to build new themes, the less files from the target system
are needed. This patch adds the option to provide a 'coreDir' parameter
that can be used to load less files from a different location than the
apps/ directory and therefore generate css files for these less files, too.