fix issue with local rpm builds
OBS injects some variables that will be available in spec files.
However, when building locally with rpmbuild, those won't be available.
Since building locally is the default use-case for this generator,
provide a working spec file for this case and add a comment on how to
use that on OBS