fix packaging sub-generator not working any longer
This is a workaround for the problem of argv not being exposed any longer
via the options object.

It is now possible to use the --package option to pass the path to a
package.json file. The old version without an name for the option can
not be supported for newer versions of yo.

Fixes #2